feat: 관리자페이지 재난제보 평가/신고 포함해서 한번에 조회 & 신고 수 long -> int

+++ core-api/src/main/java/Hongik_SafeMap_Server/domain/disaster_report/service/DisasterReportService.java
@@ -266,7 +266,7 @@ public class DisasterReportService {
         }
 
         DisasterReportEvaluation evaluation = evaluationRepository.findDisasterReportEvaluationById(disasterReportId);
-        long accusationCount = accusationRepository.countByDisasterReportId(disasterReportId);
+        int accusationCount = accusationRepository.countByDisasterReportId(disasterReportId);
 
         return DisasterReportStatisticsResponse.of(disasterReportId, evaluation, accusationCount);
     }
